refactored form handler
