+++ templates/client/forms/_register.html.twig
@@ -28,7 +28,7 @@
 
 	{% if message.success is not null %}
 		<p class="text-justify text-success mb-1" id="message">Votre compte a bien été créé.</p>
-		<p class="text-justify text-success mt-1 mb-2">Un mail de verification vous a été envoyé, veuillez suivre les indications qui y mentionnées pour vérifier votre compte.</p>
+		<p class="text-justify text-success mt-1 mb-2">Un mail de vérification vous a été envoyé, veuillez suivre les indications qui y sont mentionnées pour vérifier votre compte.</p>
 	{% endif %}
 
 	{% if message.error is not null %}
